// server/types.ts
export type TaxRegime = 'SimplesNacional' | 'LucroPresumido' | 'LucroReal';

export interface Client {
  id: number;
  name: string;
  company: string;
  email: string;
  phone: string;
  status: 'Ativo' | 'Inativo';
  taxRegime: TaxRegime;
  cnaes: string[];
  keywords: string[];
  businessDescription: string;
}
// ...
export type TaskStatus = 'Pendente' | 'Concluida';

export interface Task {
    id: number;
    clientId: number;
    description: string;
    status: TaskStatus;
    isRecurring: boolean;
    createdBy: string;
    creationDate: string;
}